requirements.txt
alembic==1.14.0
annotated-types==0.7.0
ansible==10.2.0
ansible-core==2.17.2
aiofiles==24.1.0
anyio==4.4.0
arrow==1.3.0
async-timeout==5.0.1
backports.tarfile==1.2.0
bidict==0.23.1
binaryornot==0.4.4
build==1.2.2.post1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==42.0.8
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.4
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.6.0
lazy_loader==0.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.20
oauthlib==3.2.2
packaging==23.2
pipdeptree==2.16.2
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
psutil==6.1.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.12.1
PyYAML==6.0.2
qrcode==7.4.2
readme_renderer==44.0
redis==5.2.1
reflex==0.7.0
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.32
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
resolvelib==1.0.1
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
starlette-admin==0.14.1
tabulate==0.9.0
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.2
twine==6.0.1
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
#-e git+ssh://git@github.com/craig8/volttron-installer.git@3dbd9d9cef459ea5a9f4cc138c3a4b22ee96050b#egg=volttron_installer
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
